first structural reorganization
