+++ archive/Archivio Penali - Vecchia Gara/SLA-Fornitori Nuova Gara/Recapito_Fulmine_PoSe_NuovaGara.py
