+++ core/Queries/ingestion-count-queries.json
