+++ core/Queries/reporting-alerts-config.json
