+++ core/async_gpd/Quadrature Async GPD.sql
