+++ core/async_gpd/Query Async GPD.sql
