+++ core/automazioni/SSDA-507_Sentat_maggiore_1anno_non_perfezionate.py
