+++ service_monitoring/automazioni/SSDA-347_scrittura_enti_critici.py
