+++ service_monitoring/automazioni/SSDA-476_Estrazione_Key_Client.py
