+++ service_monitoring/automazioni/SSDA-526_weekly_aggregato_non_perfezionati.py
