+++ service_monitoring/consolidatore/conformità/Specchietto_CON016_CON018.sql
