+++ service_monitoring/consolidatore/wi7/wi7_Consolidatore.py
