+++ service_monitoring/enti/key_client/KeyClient.sql
