+++ service_monitoring/enti/monitoraggio/Query_andamento.sql
