+++ service_monitoring/recapitisti/conformità/Affidi_PoSe.sql
