+++ service_monitoring/recapitisti/criticità/Estrazione_criticità.sql
