+++ service_monitoring/recapitisti/wi7/wi7_recapitista.sql
