+++ supporto_fatturazione/passivo/consolidatore/PickUp.py
