+++ supporto_fatturazione/passivo/consolidatore/Stampa_Imbustamento_Delta1.py
