+++ supporto_fatturazione/passivo/consolidatore/Stampa_Imbustamento_Delta2.py
