+++ supporto_fatturazione/passivo/recapitisti/Penale_Fulmine_nuova_gara.py
